+++ lib/appium_lib_core/common/base/bridge/mjsonwp.rb
@@ -33,8 +33,24 @@ module Appium
           include Device::TouchActions
           include Device::ExecuteDriver
 
+          attr_reader :available_commands
+
+          def initialize(capabilities, session_id, **opts)
+            @available_commands = ::Appium::Core::Commands::MJSONWP::COMMANDS.dup
+            super(capabilities, session_id, **opts)
+          end
+
           def commands(command)
-            ::Appium::Core::Commands::MJSONWP::COMMANDS[command]
+            @available_commands[command]
+          end
+
+          # command for Appium 2.0.
+          def add_command(method:, url:, name:, &block)
+            raise ::Appium::Core::Error::ArgumentError, "#{name} is already defined" if @available_commands.key? name
+
+            @available_commands[name] = [method, url]
+
+            ::Appium::Core::Device.add_endpoint_method name, &block
           end
 
           # Returns all available sessions on the Appium server instance