+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Read `release_notes.md` for commit level details.
 
 ### Deprecations
 
+## [4.7.0] - 2021-07-17
+
+### Enhancements
+- Add `satellites` option in `Appium::Core::Base::Driver#set_location`
+
 ## [4.6.0] - 2021-06-03
 
 ### Enhancements
@@ -18,7 +23,7 @@ Read `release_notes.md` for commit level details.
 ## [4.5.0] - 2021-03-14
 
 ### Enhancements
-- Add `speed` argument for `Appium::Core::Base::Driver#set_location` since Appium 1.21.0 
+- Add `speed` argument for `Appium::Core::Base::Driver#set_location` since Appium 1.21.0
 - Add `multiple` and `match_neighbour_threshold` arguments for `Appium::Core::Base::Driver#find_image_occurrence`
 
 ## [4.4.1(4.4.0)] - 2021-02-15(2021-02-13)